refactor: move methods to power-shaping.go
